requirements.txt
playwright==1.46.0
Pillow==10.3.0
ImageHash==4.3.1
requests==2.32.3
python-dateutil==2.9.0.post0
pandas==2.3.3
Flask==3.0.3
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
openpyxl==3.1.3
xlsxwriter==3.2.0
gunicorn==22.0.0
pymongo==4.8.0
APScheduler==3.10.4
